requirements.txt
Pillow>=8.0.0
PyDrive2
aiohttp>=3.8.0
asyncurban
cowpy
dnspython
emoji
gTTS>=2.2.0
gTTS-token>=1.1.3
git+https://github.com/voins/google-images-download.git@2cd68173c961324a2c41c61e8b6f40a49663ce60
gitpython
google-api-python-client
googletrans==4.0.0rc1
hachoir
humanize
lxml
oauth2client
psutil
pyDownload
pybase64>=1.0.0
pylast
pymongo[srv]>=4.0.0
python-dotenv
pytube>=12.0.0
pytz
redis>=4.0.0
requests>=2.28.0
search-engine-parser
speedtest-cli>=2.1.0
spotipy>=2.20.0
telegraph
telethon>=1.24.0
text-to-owo
wikipedia>=1.4.0
